requirements.txt
notion-client>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
